warning cleanup

+++ src/collisiondetector.cpp
@@ -24,7 +24,7 @@ void CollisionDetector::findCollisions(float elapsed_ms)
     for (size_t i = 0; i < boxes.size(); i++) {
         collisions[i].clear();
 		if (!boxes[i].removed) {
-			for (int j = 0; j < boxes.size(); j++) {
+			for (size_t j = 0; j < boxes.size(); j++) {
 				if (i != j && !boxes[j].removed) {
 					// Detect moving collisions
 					CollisionDetection::CollisionInfo collision = CollisionDetection::aabbMinkowskiCollisions(boxes[i], boxes[j], elapsed_ms);
